Moved code to a separate package AppCore
