+++ SpockOpaleye/appcore/app/Main.hs
@@ -1,0 +1,6 @@
+module Main where
+
+import Lib
+
+main :: IO ()
+main = someFunc
