+++ SpockOpaleye/appcore/appcore.cabal
@@ -1,0 +1,67 @@
+name:                appcore
+version:             0.1.0.0
+synopsis:            Initial project template from stack
+description:         Please see README.md
+homepage:            https://github.com/githubuser/appcore#readme
+license:             BSD3
+license-file:        LICENSE
+author:              Author name here
+maintainer:          example@example.com
+copyright:           2016 Author name here
+category:            Web
+build-type:          Simple
+extra-source-files:  README.md
+cabal-version:       >=1.10
+
+library
+  hs-source-dirs:      src
+                       ,src/User
+                       ,src/Role
+                       ,src/Tenant
+  exposed-modules:     AppCore
+  other-modules:       Auditable
+                       ,AuditableTH
+                       ,CryptoDef
+                       ,OpaleyeDef
+                       ,UserDefs
+                       ,TenantDefs
+                       ,RoleDefs
+                       ,Lenses
+                       ,UserId
+                       ,TenantId
+                       ,Utils
+                       ,AppM
+                       ,ApiBase
+                       ,UserApi
+                       ,TenantApi
+  build-depends:       base >= 4.7 && < 5
+                       , aeson
+                       , transformers
+                       , text
+                       , lens
+                       , time
+                       , old-time
+                       , product-profunctors
+                       , profunctors
+                       , vector
+                       , bytestring
+                       , opaleye
+                       , postgresql-simple
+                       , template-haskell
+                       , unordered-containers
+                       , bcrypt
+                       , mtl
+  default-language:    Haskell2010
+
+test-suite appcore-test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test
+  main-is:             Spec.hs
+  build-depends:       base
+                     , appcore
+  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
+  default-language:    Haskell2010
+
+source-repository head
+  type:     git
+  location: https://github.com/githubuser/appcore
