+++ SpockOpaleye/appcore/src/CryptoDef.hs
