+++ SpockOpaleye/src/Email.hs
@@ -3,10 +3,10 @@
 
 module Email where
 
+import           AppCore
 import           Network.Mail.Mime
 import           Network.Mail.SMTP  hiding (simpleMail)
 
-import           ApiBase
 import           Conf               (apikey)
 import           Control.Concurrent
 import           Control.Lens
@@ -17,10 +17,7 @@ import           Data.String.Here
 import qualified Data.Text          as T
 import qualified Data.Text.Lazy     as LT
 import qualified Data.ByteString.Lazy as L
-import           DataTypes
 import           System.FilePath (takeFileName)
-import TenantDefs
-import Lenses
 
 sendgridMail :: Mail -> IO ()
 sendgridMail mail = do
