+++ SpockOpaleye/src/UserServices.hs
@@ -2,16 +2,14 @@
 
 module UserServices where
 
+import           AppCore
 import           Control.Lens
-import           ApiBase
 import           Email
-import           DataTypes
 import qualified Data.Text                  as T
 import           Data.Monoid
 import           Control.Monad.IO.Class
 import           Validations
 import           TenantApi
-import TenantDefs
 
 doCreateTenant :: TenantIncoming -> AppM (Either T.Text Tenant) 
 doCreateTenant  incomingTenant = do
